32 Update UI for Authenticated Pages
