+++ complete-application/src/App.tsx
@@ -24,7 +24,6 @@ function App() {
         <Routes>
           <Route path="/" element={<HomePage />} />
           <Route path="/account" element={<AccountPage />} />
-          <Route path="/make-change" element={<MakeChangePage />} />
           <Route path="*" element={<Navigate to="/" />} />
         </Routes>
     </div>
