+++ complete-application/src/components/MenuBar.tsx
@@ -9,7 +9,6 @@ export default function MenuBar() {
       {
         isLoggedIn ? (
           <>
-            <NavLink to="/make-change" className="menu-link">Make Change</NavLink>
             <NavLink to="/account" className="menu-link">Account</NavLink>
           </>
         ) : (
